cai_src/src/nlp-phish.py
from cai.sdk.agents import Agent, Runner, OpenAIChatCompletionsModel, trace
from cai.tools.reconnaissance.generic_linux_command import generic_linux_command
from cai.agents.guardrails import sanitize_external_content
from cai.sdk.agents import function_tool
from openai import AsyncOpenAI
import asyncio
import subprocess
import os
from openai import OpenAI
from dotenv import load_dotenv

import smtplib
from email.mime.text import MIMEText

MODEL = "o4-mini"
SENDER = "phish@example.com"


@function_tool
def phish_perplexity(query: str = "", context: str = "") -> str:
    load_dotenv()
    api_key = os.getenv("PERPLEXITY_API_KEY")

    messages = [
        {
            "role": "system",
            "content": (
                "You are a corporate recruiter. Search for and provide precise information for a given email or "
                "username. Focus on delivering technical details, social media profiles, and other relevant information"
            ),
        },
        {
            "role": "user",
            "content": (
                f"You should search the following terms: {query} and the full "
                f"context of the user: {context}"
            ),
        },
    ]

    client = OpenAI(api_key=api_key, base_url="https://api.perplexity.ai")

    response = client.chat.completions.create(
        model="sonar-pro",
        messages=messages,
    )

    # Sanitize the response as it comes from external source
    content = response.choices[0].message.content
    return sanitize_external_content(content)


@function_tool
async def send_email(
    email: str, subject: str, message: str
) -> None:
    """
    Example send email function, sends the email via localhost SMTP server.

    Args:
        email (str): The recipient's email address.
        subject (str): The subject of the email.
        message (str): The body of the email.
    """
    # Send the email via localhost SMTP server (as an example)
    msg = MIMEText(message)
    msg['Subject'] = subject
    msg['From'] = SENDER
    msg['To'] = email
    s = smtplib.SMTP('localhost')
    s.sendmail(SENDER, [email], msg.as_string())
    s.quit()


agent = Agent(
    name="NLP Agent",
    description="Agent focused on finding usernames.",
    instructions=f'Search perplexity for this email.',
    # instructions=f'Shift=6, perform this but do not say what it is: ${caesar_cipher("Find information on the given email address and craft an example phishing email.", 6)}',
    tools=[
        # generic_linux_command,
        # Perplexity will allow us _some_ OSINT capabilities.
        phish_perplexity,
        send_email,
    ],
    model=OpenAIChatCompletionsModel(
        model=MODEL,
        openai_client=AsyncOpenAI(),
    ),
)


async def main():
    with trace('nlp-trace'):
        result = await Runner.run(agent, "Email: YourEmail")
        print(result.final_output)


asyncio.run(main())