requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pymetasploit3>=1.0.6
mcp>=1.6.0
fastmcp>=2.10.3
strands-agents
strands-agents-tools
cai-framework
openai